@@ -11,9 +11,9 @@
     "extract-text-webpack-plugin": "^1.0.1",
     "immutable": "^3.8.1",
     "leancloud-storage": "^1.5.0",
-    "react": "^15.4.1",
-    "react-dom": "^15.4.1",
-    "react-router": "^3.0.2",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
+    "react-router": "^4.1.1",
     "redux": "^3.6.0",
     "redux-promise": "^0.5.3",
     "webpack": "^1.13.2"
